requirements.txt
fastapi==0.112.2
uvicorn[standard]==0.30.6
python-dotenv==1.0.1
openai>=1.40.0
jinja2==3.1.4
